+++ lib/src/quill_editor_wrapper.dart
@@ -1,5 +1,6 @@
 import 'dart:async';
 import 'dart:convert';
+
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:quill_html_editor/quill_html_editor.dart';
@@ -202,24 +203,18 @@ class QuillHtmlEditorState extends State<QuillHtmlEditor> {
   /// This method generated the html code that is required to render the quill js editor
   /// We are rendering this html page with the help of webviewx and using the callbacks to call the quill js apis
   String _getQuillPage({required double height, required double width}) {
-    double finalHeight = width < 350
-        ? height - 110
-        : width < 600
-            ? height - 85
-            : height - 65;
-
     return '''
  <!DOCTYPE html>
       <html>
       <head>
       <meta name="viewport" content="width=device-width, initial-scale=1, minimum-scale=1, maximum-scale=1">
-      <link href="https://cdn.quilljs.com/1.3.6/quill.snow.css" rel="stylesheet" />
+      <link href="https://cdn.quilljs.com/1.3.7/quill.snow.css" rel="stylesheet" />
       <style>
       .ql-container.ql-snow {
       margin-top:0px;
       width:100%;
       border:none;
-      height: ${finalHeight.toInt() - 43}px;
+      height: ${height.toInt()}px;
       min-height:100%;
       }
       .ql-toolbar { 
@@ -276,12 +271,12 @@ class QuillHtmlEditorState extends State<QuillHtmlEditor> {
       
       <!-- Create the editor container -->
       <div style="position:relative;margin-top:0em;">
-      <div id="editorcontainer" style="height:${finalHeight.toInt()}px; min-height:100%; overflow-y:auto;margin-top:0em;">
-      <div id="editor" style="min-height:100%; height:${finalHeight.toInt() - 43}px;  width:100%;"></div>
+      <div id="editorcontainer" style="height:${height.toInt()}px; min-height:100%; overflow-y:auto;margin-top:0em;">
+      <div id="editor" style="min-height:100%; height:${height.toInt()}px;  width:100%;"></div>
       </div>
       </div>
       <!-- Include the Quill library -->
-      <script src="https://cdn.quilljs.com/1.3.6/quill.js"></script>
+      <script src="https://cdn.quilljs.com/1.3.7/quill.js"></script>
       
       <!-- Initialize Quill editor -->
       <script>
@@ -319,7 +314,7 @@ class QuillHtmlEditorState extends State<QuillHtmlEditor> {
         if(isIOS){
         element.style.height = ((height/ratio - element.offsetTop) + "px");
         }else{
-        element.style.height = ((height - element.offsetTop-60) + "px");
+        element.style.height = ((height - element.offsetTop) + "px");
         }
       
       }
@@ -333,23 +328,23 @@ class QuillHtmlEditorState extends State<QuillHtmlEditor> {
       }
       editor.applyGoogleKeyboardWorkaround = true
       editor.on('editor-change', function (eventName, ...args) {
+       
           if (eventName === 'text-change') {
             // args[0] will be delta
             var ops = args[0]['ops']
             if(ops===null){
             return
             }
-            var oldSelection = editor.getSelection()
+            var oldSelection = editor.getSelection(true)
             var oldPos = oldSelection.index
             var oldSelectionLength = oldSelection.length
             if (ops[0]["retain"] === undefined || !ops[1] || !ops[1]["insert"] || !ops[1]["insert"] ||ops[1]["list"] === "bullet"|| ops[1]["list"] === "ordered"  || ops[1]["insert"] !=  "\\n" || oldSelectionLength > 0) {
               return
             }
-          
             setTimeout(function () {
-              var newPos = editor.getSelection().index
+              var newPos = editor.getSelection(true).index
               if (newPos === oldPos) {
-                editor.setSelection(editor.getSelection().index + 1, 0)
+                editor.setSelection(editor.getSelection(true).index + 1, 0)
               }
             }, 30);
             //onRangeChanged();
@@ -388,21 +383,21 @@ class QuillHtmlEditorState extends State<QuillHtmlEditor> {
       });
         
        quilleditor.on('selection-change', function(eventName, ...args) {
-             /// console.log('selection changed');
+              // console.log('selection changed');
                onRangeChanged(); 
           });
           
       quilleditor.on('text-change', function(eventName, ...args) {
-              console.log('text changed');
+           // console.log('text changed');
              onRangeChanged(); 
              OnTextChanged(quilleditor.root.innerHTML);
           });
       
      function onRangeChanged(){
-           var range = quilleditor.getSelection();
-           var format = quilleditor.getFormat();
-           
-              if (range) {
+     try{
+    
+           var range = quilleditor.getSelection(true);   
+            if (range !=null) {
                   if (range.length == 0) {
                     var format = quilleditor.getFormat();
                      formatParser(format);
@@ -412,6 +407,9 @@ class QuillHtmlEditorState extends State<QuillHtmlEditor> {
                  }
               } else {
                   console.log('Cursor not in the editor');
+              }}
+              catch(e){
+              console.log(e);
               }
      } 
       
@@ -446,7 +444,7 @@ class QuillHtmlEditorState extends State<QuillHtmlEditor> {
       }
         
       quilleditor.root.addEventListener("focus",function (){  
-      // onRangeChanged(); 
+
       resizeElementHeight(document.getElementById("editorcontainer"),2);
       resizeElementHeight(document.getElementById("editor"),2);       
       
@@ -460,7 +458,7 @@ class QuillHtmlEditorState extends State<QuillHtmlEditor> {
       
      function getSelection()
       {
-       var range = quilleditor.getSelection();
+       var range = quilleditor.getSelection(true);
        if(range){
          return range.length;
        }
@@ -469,7 +467,7 @@ class QuillHtmlEditorState extends State<QuillHtmlEditor> {
       
    function getSelectionRange()
       {
-       var range = quilleditor.getSelection();
+       var range = quilleditor.getSelection(true);
       if(range){
          var rangeMap = {};
          rangeMap['length'] = range.length;
@@ -495,7 +493,7 @@ class QuillHtmlEditorState extends State<QuillHtmlEditor> {
       
       function embedVideo(videlUrl) 
       {  
-        var range = quilleditor.getSelection();
+        var range = quilleditor.getSelection(true);
         if(range){
           quilleditor.insertEmbed(range.index, 'video', videlUrl);
         }
@@ -504,7 +502,7 @@ class QuillHtmlEditorState extends State<QuillHtmlEditor> {
       
       function embedImage(img) 
       {  
-        var range = quilleditor.getSelection();
+        var range = quilleditor.getSelection(true);
         if(range){
           quilleditor.insertEmbed(range.index, 'image', img);
         }
@@ -520,7 +518,7 @@ class QuillHtmlEditorState extends State<QuillHtmlEditor> {
       
       function setFormat(format,value){     
       if(format == 'clean'){
-        var range = quilleditor.getSelection();
+        var range = quilleditor.getSelection(true);
           if (range) {
             if (range.length == 0) {
              quilleditor.removeFormat(range.index,quilleditor.root.innerHTML.length);
@@ -611,8 +609,12 @@ class QuillEditorController {
   }
 
   /// [getSelectionRange] to get the text selection range from editor
-  Future<dynamic> getSelectionRange() async {
-    return await _editorKey?.currentState?._getSelectionRange();
+  Future<SelectionModel> getSelectionRange() async {
+    var selection = await _editorKey?.currentState?._getSelectionRange();
+
+    return selection != null
+        ? SelectionModel.fromJson(jsonDecode(selection))
+        : SelectionModel(index: 0, length: 0);
   }
 
   /// [setSelectionRange] to select the text in the editor by index
@@ -656,5 +658,23 @@ class QuillEditorController {
   ///[dispose] dispose function to close the stream
   void dispose() {
     _changeController?.close();
+  }
+}
+
+///[SelectionModel] a model class for selection range
+class SelectionModel {
+  /// [index] index of the cursor
+  int? index;
+
+  ///[length] length of the selected value
+  int? length;
+
+  ///[SelectionModel] a model class constructor for selection range
+  SelectionModel({this.index, this.length});
+
+  ///[SelectionModel.fromJson] extension method to get selection model from json
+  SelectionModel.fromJson(Map<String, dynamic> json) {
+    index = json['index'];
+    length = json['length'];
   }
 }