+++ CHANGELOG.md
@@ -1,7 +1,12 @@
-## 2.0.0-beta.1
+## 2.0.1
 
 - Added onTextChanged listeners
-- Added custom buttons feature
+- Added ability to add custom toolbar buttons
+- Added ability to change the padding of toolbar buttons
+- Fixed bug - OnTextChanged is not firing when text is copied to editor
+- Changed getSelectionRange return type from dynamic to a selection model
+- Fixed bug - removed invisible extra bottom padding on focus
+- Other minor bug fixes